postgresql/requirements.txt
akeyless==3.3.15
backoff==2.2.1
certifi==2023.7.22
Deprecated==1.2.14
googleapis-common-protos==1.59.1
grpcio==1.51.3
grpcio-health-checking==1.43.0
importlib-metadata==6.8.0
opentelemetry-api==1.19.0
opentelemetry-distro==0.36b0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
protobuf==4.23.4
psutil==5.9.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
wrapt==1.15.0
zipp==3.16.2
akeyless==3.3.15
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
Deprecated==1.2.14
googleapis-common-protos==1.59.1
grpcio==1.51.3
grpcio-health-checking==1.43.0
idna==3.4
importlib-metadata==6.8.0
opentelemetry-api==1.19.0
opentelemetry-distro==0.36b0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-instrumentation==0.36b0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
protobuf==4.23.4
psutil==5.9.2
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
wrapt==1.15.0
zipp==3.16.2
